backend/pyproject.toml
[tool.poetry]
name = "pentest-framework"
version = "0.1.0"
description = "Web Application Penetration Testing Framework"
authors = ["You <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlmodel = "^0.0.19"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.30.0"
alembic = "^1.13.1"
pydantic = "^2.7.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
celery = "^5.4.0"
redis = "^5.0.4"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
ruff = "^0.4.8"
mypy = "^1.10.0"